Kit::JsonApi Refactor Url::Parser by moving logic to new files
